+++ src/canvas/Canvas.ts
@@ -389,8 +389,8 @@ export class Canvas extends SelectableCanvas implements CanvasOptions {
   private _onDragProgress(e: DragEvent) {
     const options = {
       e,
-      target: this._dragSource as FabricObject | undefined,
-      dragSource: this._dragSource as FabricObject | undefined,
+      target: this._dragSource,
+      dragSource: this._dragSource,
       dropTarget: this._draggedoverTarget as FabricObject,
     };
     this.fire('drag', options);